requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# LLM Client (Groq / OpenAI Compatible)
groq==0.6.0
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1

# Pydantic v2 (required by FastAPI)
pydantic==2.7.1
pydantic-core==2.18.2

# Optional utilities
requests==2.32.3
streamlit==1.37.0
websocket-client==1.8.0


# RAG Dependencies (Safe – No LangChain)
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.2

# For Async/Streaming support
anyio==4.4.0
python-multipart==0.0.9
